+++ e2e/helpers/admin-attributes-helper.js
@@ -34,6 +34,22 @@ helper.getSection = function(item) {
     };
 };
 
+helper.getTagsSection = function(item) {
+    let section = $$('.admin-attributes-section').get(item);
+
+    return {
+        el: section,
+        rows: function() {
+            return section.$$('.table-main > div');
+        }
+    };
+};
+
+
+helper.getTagsFilter = function() {
+    return $('.table-header .e2e-tags-filter');
+}
+
 helper.getStatusNames = function(section) {
     return section.$$('.status-name span').getText();
 };
@@ -93,6 +109,14 @@ helper.getGenericForm = function(form) {
     obj.name = function() {
         return form.$('.status-name input');
     };
+
+    obj.colorBox = function() {
+        return form.$('.edition .current-color');
+    }
+
+    obj.colorText = function() {
+        return form.$('.select-color input');
+    }
 
     return obj;
 };