+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add the tribe button to link stories from tree.taiga.io with gigs in tribe.taiga.io.
 - Errors (not found, server error, permissions and blocked project) don't change the current url.
 - Attachments image slider
+- New admin area to edit the tag colors used in your project
 
 ### Misc
 - Lots of small and not so small bugfixes.